+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Page Up / Page Down (and Shift variants) page the member nick list when it
+  has keyboard focus. Home / End jump to the first or last nick. Transcript
+  paging is unchanged while the composer is focused.
+
 ## [1.0.1] - 2026-09-24
 
 ### Fixed